requirements.txt
temporalio==1.6.0
langchain==0.2.0
langchain-openai==0.1.7
langgraph==0.0.60
pydantic>=2.9.0
pydantic-settings>=2.5.2
python-dotenv==1.0.1
mcp>=1.2.0
